Implemented raising an audio notification when a given timer elapses.
At this time, a single audio notification is raised when a given timer elapses. The current implementation is a bit fragile as it requires the expected audio file to be present alongside the executable. I would like to make the notification customizable at some point in the future. In addition, I think the audio notification should be repeatedly played until the user acknowledges that the timer ran out.